latest video code
Jaime, I am stuck with the error 'positional argument missing, time lock' for the snipper function. I have tried defining timelock as an arbitrary number but that doesn't make a difference to the error. Should the value(s) of timelock be stored in the rats.pickle file? Unfortunately I can't view this on my Mac so can't check. please Help!